pyproject.toml
[project]
name = "python-fp-lint"
version = "0.1.0"
description = "Functional-programming lint rules for Python — Semgrep + ast-grep + beniget"
requires-python = ">=3.10"
dependencies = [
    "beniget>=0.5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
